feat: create config file for reservation preferences
