+++ build.sbt
@@ -14,6 +14,7 @@ val root = Project("resy-booking-bot", file("."))
     scalacOptions += "-Ywarn-unused",
     libraryDependencies ++= Seq(
       "com.typesafe.play"        %% "play-ahc-ws"     % "2.8.16",
+      "com.github.pureconfig"    %% "pureconfig"      % "0.17.1",
       "org.apache.logging.log4j" %% "log4j-api-scala" % "12.0",
       "org.apache.logging.log4j"  % "log4j-core"      % "2.13.0" % Runtime,
       "org.scalatest"            %% "scalatest"       % "3.2.12" % Test,
