+++ src/main/resources/resyConfig.conf
@@ -1,0 +1,49 @@
+############
+# ResyKeys #
+############
+# Your user profile API key which can be found via your browser web console in your headers called "authorization"
+# e.g.
+# resyKeys.api-key="MY_API_KEY"
+resyKeys.api-key=
+# Your user profile authentication token which can be found via your browser web console in your headers called
+# "x-resy-auth-token"
+# e.g.
+# resyKeys.auth-token="MY_AUTH_TOKEN"
+resyKeys.auth-token=
+
+######################
+# ReservationDetails #
+######################
+# Date of the reservation in YYYY-MM-DD format
+# e.g.
+# resDetails.date="2099-01-30"
+resDetails.date=
+# Size of the party reservation
+# e.g.
+# resDetails.party-size=2
+resDetails.party-size=
+# Unique identifier of the restaurant where you want to make the reservation
+# e.g.
+# resDetails.venue-id=123
+resDetails.venue-id=
+# Priority list of reservation times and table types. Time is in military time HH:MM:SS format. If no preference on
+# table type, then simply don't set it.
+# e.g.
+# resDetails.res-time-types=[
+#     {reservation-time="18:00:00", table-type="Dining Room"},
+#     {reservation-time="18:00:00", table-type="Patio"},
+#     {reservation-time="18:15:00"}
+# ]
+resDetails.res-time-types=
+
+#############
+# SnipeTime #
+#############
+# Hour of the day when reservations become available and when you want to snipe
+# e.g.
+# snipeTime.hours=9
+snipeTime.hours=
+# Minute of the day when reservations become available and when you want to snipe
+# e.g.
+# snipeTime.minutes=0
+snipeTime.minutes=
